# opt: use faster NEON bitmask and fix skip number bug