--- a/src/util/simd/mod.rs
+++ b/src/util/simd/mod.rs
@@ -1,3 +1,4 @@
+mod bits;
 mod traits;
 
 #[doc(hidden)]